requirements.txt
Pillow
pillow-heif
numpy
qrcode
reportlab
Flask
requests
black